Ensemble Number, Documentation
Reading through the user manual I found there was another byte in the variable leader data for rollover. This was required to make it so numbering was correct - would have had a lot of ensembles with the same number. Required changing the data type for number to accommodate this
Added comments laying out which bytes correspond to what data to make it easier to track what's going on

+++ decode_bin.py
@@ -155,7 +155,7 @@ def config_file(fname: str, n_cells: int, cfg: Config):
         ens_group = f.create_group("ensembles")
 
         ds = ens_group.create_dataset("number", shape=(
-            0,), dtype="uint16", chunks=batch_size, maxshape=(None,))
+            0,), dtype="uint32", chunks=batch_size, maxshape=(None,))
         add_metadata(ds, units="1", long_name="Ensemble Number")
 
         ds = ens_group.create_dataset("mtime", shape=(
